pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.1",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "whitenoise>=6.6.0",
    "openpyxl>=3.1.5",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.40.0",
]
